+++ frontend/src/hooks/food/useInfiniteFoodListScroll.ts
@@ -1,32 +1,8 @@
-import { useCallback, useEffect, useRef, useState } from 'react';
+import { useCallback, useEffect, useRef } from 'react';
 
-import { initialSelectedFilterList } from '@/context/food';
+import useValidQueryString from '@/hooks/common/useValidQueryString';
+import { useFoodListInfiniteQuery } from '@/hooks/query/food';
 import type { KeywordEn } from '@/types/food/client';
-import { parseCheckList } from '@/utils/parseCheckList';
-
-import useValidQueryString from './common/useValidQueryString';
-import { useFoodListInfiniteQuery } from './query/food';
-
-export const useFoodListFilter = () => {
-  const [selectedFilterList, setSelectedFilterList] = useState(initialSelectedFilterList);
-
-  const parsedSelectedFilterList = parseCheckList(selectedFilterList);
-
-  const toggleFilter = (keyword: KeywordEn, filter: string) => {
-    const targetFilterList = structuredClone(selectedFilterList)[keyword];
-    const selected = targetFilterList.has(filter);
-
-    selected ? targetFilterList.delete(filter) : targetFilterList.add(filter);
-
-    setSelectedFilterList(prev => ({ ...prev, [keyword]: new Set(targetFilterList) }));
-  };
-
-  const resetSelectedFilterList = () => {
-    setSelectedFilterList(initialSelectedFilterList);
-  };
-
-  return { selectedFilterList, parsedSelectedFilterList, toggleFilter, resetSelectedFilterList };
-};
 
 export const useInfiniteFoodListScroll = () => {
   const queries = useValidQueryString<KeywordEn>([
